requirements.txt
anyio
contextlib2
contextvars
fastapi
idna
immutables
sniffio
starlette
typing-extensions
pymongo
python-dotenv
uvicorn
bcrypt

passlib
python-jose
email-validator
requests-async